Move context to providers/ProjectLayoutProvider
Extract context, provider component, and hook into a dedicated
providers folder using proper React context pattern instead of
Outlet context.
