--- src/frontend/lib/projectLayoutContext.ts
@@ -1,13 +1,0 @@
-import { useOutletContext } from "react-router-dom";
-
-export interface ProjectLayoutContext {
-  projectId: string | undefined;
-  sidebarOpen: boolean;
-  setSidebarOpen: (open: boolean) => void;
-  onNewAgent: () => void;
-  onBrowseCatalog: () => void;
-}
-
-export function useProjectLayout() {
-  return useOutletContext<ProjectLayoutContext>();
-}
